+++ http/src/main/java/com/mx/path/connect/http/certificate/FieldSettingsValidationError.java
@@ -1,4 +1,4 @@
-package com.mx.path.connect.http.certificates;
+package com.mx.path.connect.http.certificate;
 
 import java.util.List;
