+++ http/src/test/groovy/com/mx/path/api/connect/http/certificate/MutualAuthProviderFactoryTest.groovy
@@ -1,8 +1,8 @@
-package com.mx.path.api.connect.http.certificates
+package com.mx.path.api.connect.http.certificate
 
-import com.mx.path.connect.http.certificates.FieldSettingsValidationError
-import com.mx.path.connect.http.certificates.MutualAuthProviderFactory
-import com.mx.path.connect.http.certificates.MutualAuthProviderKeystore
+import com.mx.path.connect.http.certificate.FieldSettingsValidationError
+import com.mx.path.connect.http.certificate.MutualAuthProviderFactory
+import com.mx.path.connect.http.certificate.MutualAuthProviderKeystore
 import com.mx.path.core.common.connect.ConnectionSettings
 import com.mx.path.core.common.connect.RequestFilter
